Fixed an issue with the char isc server
We had an issue where the char isc server was using the wrong varible to check to see if the isc server was the login server. This caused an issue where we would never tell the login server about a map server and would just go in an infinite loop recv'ing the same packet.