Fix compiler error on missing sql header files
Made a change to how we generate the sql header files. Core now depends on a target that will force the sql header generator script to run before building core. We want this to happen to prevent issues where when we try to build the precompiled header, the file is missing.