Fixed a few bugs
Fixed a bug where if you picked a respawn option that the current map doesn't have, it would place you at 0,0.
Increased the visibility range from 10 units to 100 units.
You should now see the attacking animation as you load an entity that currently has a target that isn't an NPC